Move local command CTA to chat header
