+++ docs/app/chat/chat-page.module.css
@@ -49,6 +49,45 @@
   outline-offset: 3px;
 }
 
+.startLocallyButton {
+  display: inline-flex;
+  height: 2.5rem;
+  min-width: 18.75rem;
+  align-items: center;
+  justify-content: center;
+  gap: 0.45rem;
+  padding-inline: 0.875rem;
+  border: 1px solid rgba(255, 255, 255, 0.14);
+  border-radius: 999px;
+  background: #0a0a0a;
+  color: #fff;
+  box-shadow:
+    inset 0 7px 12px -6px rgba(255, 255, 255, 0.16),
+    inset 0 -5px 8px -2px rgba(0, 0, 0, 0.6);
+  font-family: var(--font-geist-mono), monospace;
+  font-size: 0.75rem;
+  font-weight: 500;
+  line-height: 1;
+  white-space: nowrap;
+  cursor: pointer;
+  transition:
+    transform 0.2s ease,
+    box-shadow 0.2s ease;
+}
+
+.startLocallyButton:hover {
+  transform: scale(0.99);
+}
+
+.startLocallyButton:focus-visible {
+  outline: 2px solid var(--openui-text-neutral-primary);
+  outline-offset: 2px;
+}
+
+.startLocallyButton svg {
+  flex-shrink: 0;
+}
+
 .modeControl {
   min-width: 0;
   margin-left: auto;
@@ -442,6 +481,12 @@
   }
 }
 
+@media (max-width: 820px) {
+  .startLocallyButton {
+    display: none;
+  }
+}
+
 @media (max-width: 760px) {
   .header {
     padding: 8px 12px;
@@ -496,6 +541,10 @@
 }
 
 @media (prefers-reduced-motion: reduce) {
+  .startLocallyButton {
+    transition: none;
+  }
+
   .loadingIndicator,
   .dialogOverlay,
   .dialogContent {
